Configuration command (#9)
* Add ConfigCommand + readme update

* too much thing

* Update README

* ConfigDB: populate table + add config on guild join & add values for Join/Leave messages

* Reduce jar size & fixed a repo link

* Update JDA version

* more config & finished config check/use

* Reformulate README

* Documented (Mod)Command constructor + added constructors for Subcommand & Subcommand Group

* Added temporary ConfigCommand options

* Finalized temporary Config implementation

* Forgot to config-ify the accept button

+++ src/joseta/database/MarkovMessagesDatabase.java
@@ -1,7 +1,9 @@
 package joseta.database;
 
 import joseta.*;
+import joseta.database.ConfigDatabase.*;
 
+import arc.files.*;
 import arc.struct.*;
 
 import net.dv8tion.jda.api.entities.*;
@@ -19,18 +21,16 @@ import java.util.stream.*;
 public class MarkovMessagesDatabase extends ListenerAdapter {
     private static final String dbFileName =  "resources/database/markov_messages.db";
     private static Connection conn;
-    private static Seq<Long> channelBlackList = Seq.with(1219013018873233512L, 1219013344099303576L, 1306634181119315979L, 1256989659448348673L);
-    private static Seq<Long> categoryBlackList = Seq.with(1219273667012460664L);
 
     private static final Pattern NO_URL_PATTERN = Pattern.compile("(https?://\\S+|www\\.\\S+[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}\\S*)");
     private static final Pattern CLEAN_COPY_PATTERN = Pattern.compile("[^a-z0-9.?!,;\\-()~\"'&$€£ \\]\\[àáâãäåæçèéêëìíîïñòóôõöùúûüýÿ ]+", Pattern.CASE_INSENSITIVE | Pattern.UNICODE_CASE);
     private static final Pattern SPACED_PATTERN = Pattern.compile("\\s+");
 
     public static void initialize() {
-        File dbFile = new File(dbFileName);
+        Fi dbFile = new Fi(dbFileName);
         try {
             if (!dbFile.exists()) {
-                dbFile.createNewFile();
+                dbFile.write().close();
                 
                 conn = DriverManager.getConnection("jdbc:sqlite:" + dbFileName);
 
@@ -57,15 +57,18 @@ public class MarkovMessagesDatabase extends ListenerAdapter {
 
         Statement stmt = conn.createStatement();
         stmt.execute(messageTable);
+        stmt.close();
     }
 
     private static void populateNewTable() {
         int count = 0;
         JosetaBot.logger.debug("Populating the Markov Messages Database...");
         for (Guild guild : JosetaBot.bot.getGuilds()) {
+            ConfigEntry config = ConfigDatabase.getConfig(guild.getIdLong());
+
             for (TextChannel channel : guild.getTextChannels()) {
-                if (channelBlackList.contains(channel.getIdLong())) continue;
-                if (channel.isNSFW() || categoryBlackList.contains(channel.getParentCategoryIdLong())) continue;
+                if (config.markovChannelBlackList.contains(channel.getIdLong())) continue;
+                if (channel.isNSFW() || config.markovCategoryBlackList.contains(channel.getParentCategoryIdLong())) continue;
 
                 for (ThreadChannel thread : channel.getThreadChannels()) count += addChannelMessageHistory(thread, guild);
                 count += addChannelMessageHistory(channel, guild);
@@ -77,7 +80,7 @@ public class MarkovMessagesDatabase extends ListenerAdapter {
     }
 
     private static int addChannelMessageHistory(GuildMessageChannel channel, Guild guild) {
-        int count = 0;        
+        int count = 0;
         try {
             for (Message message : channel.getIterableHistory().takeAsync(10000).thenApply(list -> list.stream().collect(Collectors.toList())).get()) {                
                 long authorId = message.getAuthor().getIdLong();
@@ -157,11 +160,13 @@ public class MarkovMessagesDatabase extends ListenerAdapter {
         long id = message.getIdLong();
         long guildId = guild.getIdLong();
         long channelId = channel.getIdLong();
+
+        ConfigEntry config = ConfigDatabase.getConfig(guildId);
         
         if (message.getAuthor().isBot() || message.getAuthor().isSystem()) return;
-        if (channelBlackList.contains(channelId)) return;
+        if (config.markovChannelBlackList.contains(channelId)) return;
         if (channel instanceof TextChannel textChannel &&
-            (textChannel.isNSFW() || categoryBlackList.contains(textChannel.getParentCategoryIdLong()))) return;
+            (textChannel.isNSFW() || config.markovCategoryBlackList.contains(textChannel.getParentCategoryIdLong()))) return;
 
         
         try (PreparedStatement pstmt = conn.prepareStatement("INSERT INTO messages "
@@ -181,6 +186,7 @@ public class MarkovMessagesDatabase extends ListenerAdapter {
     }
 
     public static void updateMessage(long id, long guildId, long channelId, String content) {
+        if (getMessageEntry(id, guildId, channelId) == null) return; // Check if message exists
         try (PreparedStatement pstmt = conn.prepareStatement("UPDATE messages SET content = ? "
                                                            + "WHERE id = ? AND guildId = ? AND channelId = ?")) {
             pstmt.setString(1, cleanMessage(content));
@@ -203,6 +209,7 @@ public class MarkovMessagesDatabase extends ListenerAdapter {
     }
 
     public static void deleteMessage(long id, long guildId, long channelId) {
+        if (getMessageEntry(id, guildId, channelId) == null) return; // Check if message exists
         try (PreparedStatement pstmt = conn.prepareStatement("DELETE FROM messages WHERE id = ? AND guildId = ? AND channelId = ?")) {
             pstmt.setLong(1, id);
             pstmt.setLong(2, guildId);
@@ -230,6 +237,7 @@ public class MarkovMessagesDatabase extends ListenerAdapter {
                     rs.getString("timestamp")
                 );
             }
+            pstmt.close();
         } catch (SQLException e) {
             JosetaBot.logger.error("Could not retrieve a message.", e);
         }
